PCHR-2339: Renaming, add helper to get user webform submissions
